Homologous gene distribution of H. Henanensis sp. nov. in 3376 species.
Note:
The values in the table represent the homologues gene number in a given species.
In this table, each row represents a species and each column represents a gene of H. Henanensis sp. nov. The order of species and genes is the same as shown in Fig. 3A.
H. Henanensis sp. nov. itself and the gene age grade are shown at the bottom two rows.